requirements.txt
python-telegram-bot==13.7
python-dotenv==0.19.2
notion-client==1.0.0
requests>=2.28.0,<3.0.0  # 更新为满足 conda 要求的版本
google-generativeai>=0.3.0
schedule>=1.1.0
python-dateutil>=2.8.2
httpx[socks]
urllib3<2.0.0  # 与 requests 2.28 兼容的 urllib3 版本
pytz==2023.3
beautifulsoup4==4.10.0
pypdf>=3.0.0
openai>=0.27.0
# pymupdf>=1.18.0  # 添加 PyMuPDF 用于 PDF 处理
# 其他项目依赖...
